requirements.txt
Flask==3.0.0
python-docx==1.1.0
pdfplumber==0.10.3
google-generativeai==0.3.2
python-dotenv==1.0.0